RND-40: Rename pom.xml properties, project packages and document taken steps
