+++ pom.xml
@@ -8,12 +8,12 @@
         <version>3.0.6</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>com.intellias</groupId> <!-- change it -->
-    <artifactId>basic-sandbox</artifactId> <!-- change it -->
+    <groupId>com.intellias</groupId>
+    <artifactId>intellias-parking</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>basic-sandbox</name> <!-- change it -->
-    <description>Basic service to start your own by copying and adjusting for your needs
-    </description> <!-- change it -->
+    <name>intellias-parking</name>
+    <description>Test project that has been created using the Java Sandbox project
+    </description>
     <properties>
         <java.version>17</java.version>
         <mapstruct.version>1.5.5.Final</mapstruct.version>
