--- src/main/java/com/intellias/basicsandbox/Application.java
@@ -1,12 +1,0 @@
-package com.intellias.basicsandbox;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(final String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
