--- src/main/java/com/intellias/basicsandbox/config/aspect/PersistenceConfig.java
@@ -1,9 +1,0 @@
-package com.intellias.basicsandbox.config.aspect;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@Configuration
-@EnableTransactionManagement
-public class PersistenceConfig {
-}
