--- src/main/java/com/intellias/basicsandbox/controller/dto/ErrorDTO.java
@@ -1,10 +1,0 @@
-package com.intellias.basicsandbox.controller.dto;
-
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-@Data
-public class ErrorDTO {
-    private HttpStatus status;
-    private String message;
-}
