--- src/main/java/com/intellias/basicsandbox/controller/exception/RestExceptionHandler.java
@@ -1,24 +1,0 @@
-package com.intellias.basicsandbox.controller.exception;
-
-import com.intellias.basicsandbox.controller.dto.ErrorDTO;
-import com.intellias.basicsandbox.service.exception.ItemNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-
-@RestControllerAdvice
-public class RestExceptionHandler {
-
-    @ExceptionHandler(ItemNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorDTO handleItemNotFound(ItemNotFoundException ex) {
-        ErrorDTO errorDTO = new ErrorDTO();
-        errorDTO.setMessage(ex.getMessage());
-        errorDTO.setStatus(NOT_FOUND);
-
-        return errorDTO;
-    }
-}
