--- src/main/java/com/intellias/basicsandbox/persistence/converter/EncryptedStringConverter.java
@@ -1,60 +1,0 @@
-package com.intellias.basicsandbox.persistence.converter;
-
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-import lombok.SneakyThrows;
-
-import javax.crypto.BadPaddingException;
-import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.spec.SecretKeySpec;
-import java.security.InvalidKeyException;
-import java.security.Key;
-import java.util.Base64;
-
-@Converter
-public final class EncryptedStringConverter implements AttributeConverter<String, String> {
-
-    private static final String AES = "AES";
-
-    private static final String SECRET = "secret-key-12345";
-
-    private final Key key;
-    private final Cipher cipher;
-
-    public EncryptedStringConverter() throws Exception {
-        key = new SecretKeySpec(SECRET.getBytes(), AES);
-        cipher = Cipher.getInstance(AES);
-    }
-
-    @SneakyThrows
-    @Override
-    public String convertToDatabaseColumn(String attribute) {
-        if (attribute == null || attribute.isBlank()) {
-            return null;
-        }
-
-        try {
-            cipher.init(Cipher.ENCRYPT_MODE, key);
-            return Base64.getEncoder().encodeToString(cipher.doFinal(attribute.getBytes()));
-        } catch (IllegalBlockSizeException | BadPaddingException | InvalidKeyException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    @Override
-    public String convertToEntityAttribute(String dbData) {
-        if (dbData == null || dbData.isBlank()) {
-            return null;
-        }
-
-        try {
-            cipher.init(Cipher.DECRYPT_MODE, key);
-            return new String(cipher.doFinal(Base64.getDecoder().decode(dbData)));
-        } catch (InvalidKeyException | BadPaddingException | IllegalBlockSizeException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-
-}
