+++ src/main/java/com/intellias/parking/config/aspect/PersistenceConfig.java
@@ -1,0 +1,9 @@
+package com.intellias.parking.config.aspect;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@Configuration
+@EnableTransactionManagement
+public class PersistenceConfig {
+}
