+++ src/main/java/com/intellias/parking/controller/dto/ErrorDTO.java
@@ -1,0 +1,10 @@
+package com.intellias.parking.controller.dto;
+
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+
+@Data
+public class ErrorDTO {
+    private HttpStatus status;
+    private String message;
+}
