+++ src/main/java/com/intellias/parking/controller/exception/RestExceptionHandler.java
@@ -1,0 +1,24 @@
+package com.intellias.parking.controller.exception;
+
+import com.intellias.parking.controller.dto.ErrorDTO;
+import com.intellias.parking.service.exception.ItemNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
+@RestControllerAdvice
+public class RestExceptionHandler {
+
+    @ExceptionHandler(ItemNotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public ErrorDTO handleItemNotFound(ItemNotFoundException ex) {
+        ErrorDTO errorDTO = new ErrorDTO();
+        errorDTO.setMessage(ex.getMessage());
+        errorDTO.setStatus(NOT_FOUND);
+
+        return errorDTO;
+    }
+}
