+++ src/main/java/com/intellias/parking/service/exception/ItemNotFoundException.java
@@ -1,0 +1,11 @@
+package com.intellias.parking.service.exception;
+
+import java.util.UUID;
+
+public class ItemNotFoundException extends RuntimeException {
+
+    public ItemNotFoundException(UUID id) {
+        super("An item with ID " + id + " not found.");
+    }
+
+}
