--- src/test/java/com/intellias/basicsandbox/utils/TestUtils.java
@@ -1,15 +1,0 @@
-package com.intellias.basicsandbox.utils;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class TestUtils {
-
-    public static String asJsonString(final Object obj) {
-        try {
-            return new ObjectMapper().writeValueAsString(obj);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-}
