+++ src/test/java/com/intellias/parking/utils/TestUtils.java
@@ -1,0 +1,15 @@
+package com.intellias.parking.utils;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class TestUtils {
+
+    public static String asJsonString(final Object obj) {
+        try {
+            return new ObjectMapper().writeValueAsString(obj);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+}
